Added two flags to tcc:
   -vt  Dumps verbose information about each step in the
           toolchain, including the argv given to each stage.

   -ve  Dumps very verbose information about the environment
           used in each step of the toolchain.  This is useful
           if -y options are to be read from environment
           variables; otherwise, it's very verbose.

In summary, these flags are useful for very low-level debugging,
mostly with changes in the toolchain, and in developing a bootstrap.
We'll have to discuss whether we need to also patch the 4.1.2 branch
in cvs.

In the meantime, there's also a diff of these changes for the FreeBSD
port version of TenDRA, 4.1.2, at:

  http://www.tendra.org/~nonce/tendra/toolchain/toolchain.diff