Significant change to tccenv(5) file handling.  This patch performs a small clean up of src/tools/tcc, and also significantly changes the parsing of tccenv files.  Whereas before, only the TENDRA_* env settings were treated as variables in -Y designated env files (e.g., '+TDFC "<TENDRA_BINDIR>/tdfc2"'), now *any* expression can be treated as a variable.  User can declare new variables with the usual '+' operation, and prepend or append using '<' and '>' as before.  For example, this:
  +TDFC                   "<TENDRA_BINDIR>/tdfc2"
  +TDFCPP                 "<TENDRA_BINDIR>/tdfc2 -E"

can now be rewritten as this:

   +TDFC                   "<TENDRA_BINDIR>/tdfc2"
   +TDFCPP                 "<TDFC> -E"

   If the variable name matches one of the tccenv(5) variables, it is
executed in the normal manner.  Thus, all existing -Y env files will
work as before.  If the variable name does not match a know tccenv(5)
variable, it is stored, and may be used in other env files.  (Thus,
the order of files given in -Y arguments may play a role.) This allows
us to completely remove all hard-coding of paths (and other
expressions) in env files.

   Parsing of env files is also improved, using in-place string
manipulation, effectively reducing the copying from 3x per line to 1x
per line.  Storage of declared variables is done via a hashtable with
pluggable hash function, should the mix of tccenv variables change
enough to create too many collisions.  At present, the size of the
table, key used, and hash function are tuned to the variables listed
in tccenv(5).

    This is a first cut, and changes may follow.  At present, tcc
works as before, but has a new capability we're going to exploit.