Bug fix.  We now detect if a user invokes tcc without providing any -Y env arguments.  Thanks to verm for finding this.
   Since missing -Y env files is an obvious user mistake, we also
issue a warning, regardless of the verbose level setting.  Under the
old code, the only hint was the internal error from the execute stage:
"The tool 'C_producer' is not available"--an unhelpful message to new
users.  Now, the user is warned that tcc was "not invoked with any -Y
env files".

   Future work must address how this warning interacts with -q and -v
flags.