An unlucky combination of function inlining and the lack of a return value can trigger an internal error in the 80x86 translator:
  $ cat unexpected_size.c
  void a(void);
  int b(void) { a(); }
  void c(void) { if (b()); }
  $ tcc -c unexpected_size.c
  trans:/tmp/tccEiQ0XO/_tcc.t: internal error: unexpected size
  tcc: Error: Caught signal 11 in
  '/usr/local/lib/TenDRA/5.0.0/FreeBSD/5.2-BETA/80x86/bin/trans'.

The reason is that one of the operands of a test_tag exp (corresponding to
the if in c()) is a top_tag exp (an empty computation) due to the missing
return expression in b().  The function cmp() cannot handle that case
which leads to the internal error.  Just return and don't emit any code in
cmp() as the value is indeterminate anyway.

The segmentation fault seen above is an independent error (still to be
fixed) caused by the fact that parts of the installer falsely assume calling
failer() terminates the translator's execution.

If you're interested: The error was triggered originally by lex-generated
code in the aicasm tool which is used to build a few FreeBSD kernel drivers.