docbook-to-man gets horribly confused about <informaltable>, so change these to <variablelist> which it understands.
Unfortunatly, it also dosn't understand <envar>, so use <filename> in it's place
to get the proper bolding we want in the manpage.

If at some point we're doing too many hacks like this, we'll just import
docbook-to-man into contrib/ and modify it accordingly, but at this point
it's just not worth it (yet).