Completely revamp our contributors list, currently there are only a few 'placeholders'.
This uses a combination of seperate files and a couple awk scripts.
Instead of having to worry about formatting the SGML (let alone in
alphabetical order) this takes care of everything for us.  Atleast this
way we'll be far more inclined to update it.

As always, if we've left anyone out, or if anyone wants to be added
send ane mail over to the docs mailing list.

Happy New Year!