Add the options map. This is mainly to help with command line option coordination and identifying undocumented options and similar things.
Please note that this, as well as much of the further command line
option coordination work that is being discussed at the moment and will
be presented concretely soon, could not possibly have been achieved so
smoothly without the help of Jesper Louis Andersen <jlouis@mongers.org>.

His SML-based TenDRA Option Collision Checker software enabled us to get
a clear overview of options and their purposes, and without it we surely
could not have progressed as quickly as we did. Once again, thanks to
Jesper and his hard work on this task.