Add trans manpages, some notes:
While this does seem rather complex, it's not.  It's easier to add similar
documentation in a central point, people are more likely to update it.  It
*always* happens... you have documentation in 5 different places, eventually
someone is going to forget to update a block and you have out dated
documentation everywhere.

These manpages are constructed mainly from a central location, common/* with
specific options placed within each manpage.sgml.  It makes life much, much
easier.