Add an entity for &trans;
Since this is just a reference document I wasn't about to do some pointless
DSSSL to dynamically swap the name based on the name of the <refentry> blocks.
So for any instances of &trans; we swap in "trans"

It's the same thing and not hard to know that it's talking about the current
translator, the only place this really counts is in the manpages, where it's
done properly.