Instead of installing into ${INSTALL_DIR}/${VER_TENDRA}/${MACH_OS}/${MACH_VERS}/${MACH_CPU} we now install into ${INSTALL_DIR} (default is ${PREFIX}/lib/TenDRA) for our toolchain files/programs.
Having the layered version is very annoying, and only useful for development.
Originally the plan was to do this for when TenDRA is a cross-compiler, however
that is a long way away yet.  For now only developers will be interested in
preserving historical versions of tendra, so they can be easily invoked for
testing purposes.

There is a new option in config.mk USE_PATH_DEV, if you set this to yes it will
go back to the previous installation path.