Fix a bug introduced in r1.10 (by me)
At the time, this was the best method, however i've done a major re-write of
the build system in how it works.  It's much, much easier to understand and creates
config files that are more human-readable.  Having two different ways to define
variables was a big mistake.

For now this is just a patch, I removed the SED_OPT_COND from the sed parameter list

PR:	33