Comment the sid transformation suffix rule, until I devise a way to make this work when someone doesn't have sid yet installed and make wants to badly treat the resultant file as out of date.
Add explicit OBJ_SDIR placement to the generated output files of the
compilation and linking phase.

Make sure to pass along the MAKEOBJDIR environment variable to the
build-prog and link-prog targets in order to make sure they look at the
right place for their object files.

Make sure the cleaning phase checks if PROG is defined in order to
not just try to delete .core and core in the parent directory.

svn path=/tendra/trunk/; revision=80