Command line option coordination.
Moved -h to -a. Though this sets the "helpflag" variable, it has nothing
                to do with help, but instead controls vertical alignment
                spacing for disp output.
Moved -V to -r. This tells disp not to rationalise the TDF it reads.
Moved -U to -l. Prints usage information associated with external tokens
                and tags, as well as linkage information within the
                capsule.
Moved -v to -V. Report version.
Moved -p to -v. Verbose mode.