Teach the producers how to print dependency information usable for make.
For each processed source file the object file depends on the source file and
the headers included from the latter.  Passing -iM to the producers causes all
headers to be printed, -iMM only those between double quotes.  The default
suffix for object files is .o, this can be overridden with -iO<.ext>.

At the moment you have to pass -Wp,-iM -E (-Wg,-iM -E for C++ files) to tcc to
use this feature, hopefully a more comfortable option will be added to tcc
later.

Reviewed by:	bp (slightly older patch)