Manual manpage update.
I'm going to write a small Makefile target to automatically update the manpages
from the mdoc dir in the doc repo.  In order to do that I needed the manpages
in the repo to be in the 'final' format.

The Crown copyright has been replaced with the BSDL.  Some old wording still
does exist but not for long.

Add two missing manpages: ppc601.aix.trans.1 680x0-sunos.trans.1