Implement C99's flexible array members.  It means that you are allowed to omit the size of an array member, provided it is the last member and there is at least one additional member.  The size of such a struct is defined as the offset to the last member.  The intended usage of this feature is to be able to allocate a struct that contains a dynamically sized array with just one allocation call.
The pragma option "flex_array_member" controls whether flexible array members
are allowed.

There is one open issue with the current implementation.  While you are not
allowed to have an array of structs with flexible array members, there is no
constraint which forbids to increment a pointer to such a struct object by one.
Since the standard describes pointer arithmetics in terms of arrays, it seems
reasonable that those expressions are undefined.  Currently the pointer is
incremented by the size of the struct, rounded up according to the installer's
idea of its alignment requirements.  This means that the expression
(char *)(&a + 1) == (char *)&a + sizeof(a) isn't guaranteed to be true if a's
type contains a flexible array member.