Fix the compilation of the following code:
  int a(void) {
	  struct b;
	  static struct b *b;
	  return b != 0;
  }

which results in
  Internal error: 'tag 1' used but not defined.

Let me first give you some background information about the declaration and
definition of variables in the producers.  After a declarator has been read,
make_object_decl() is called for objects.  There, the function
search_member() is called which adds the identifier of the object just read
to the current namespace.  The 'last' field of the namespace is set to point
to this new namespace member.  Later, add_compound_stmt() checks to see if
'last' has a newer value than 'prev' (unchanged by search_member()) and
calls make_decl_stmt() to create the new variable.

A namespace member has two entries, 'id' for an object name and 'alt' for a
tag name.  If search_member() finds a match in either of these entries, it
returns this member (the rules are slightly different for C++ but the bug
exists there too).

Now the problem with the above code is following.  Line 2 declares struct b,
this adds an entry to the namespace of function a with 'id' = 0 and 'alt' =
b.  For line three, make_object_decl() calls search_member() which finds
this member.  'id' is made to point to the object b and 'alt' continues to
point to struct b.  But when add_compound_stmt() executes, 'last' still
equals 'prev' because search_member() didn't create a new member.  Thus
make_decl_stmt() is not called and the variable b is not created which leads
to the internal error.

The function update_member() exists to work around the problem.  It zeroes
out the old member for b and creates a new one at the end.  The field 'last'
then points to that member and add_compound_stmt() works as wanted.
Unfortunately update_member() isn't called for static objects.  To be
honest, I have no idea why.  This commit removes the check for automatic
storage duration and calls update_member() also for objects with static
storage duration (extern is handled a few lines above this check).

Bugzilla:	3