Add support for the combination -Yc99 and -Xmode.  This commit adjusts the environment files to use the new Xmode files in startup/Modes.c99.
The environment variable MODES_DIR is added to point to the directory
startup/Modes or startup/Modes.c99, depending on whether -Yc99 was specified
or not.  The environment Xmode then uses this variable to set the STARTUP
variable to the corresponding mode file.  For consistency and future needs I
did that for C++ too (MODES_CPP_DIR), though at the moment startup/Modes.cpp
is the only choice.