Relax the restrictions on '#pragma weak id1 = id2'.  Previously we required both id1 and id2 to be declared, with this change this is not needed for id1 anymore.  Now the producers act as if they have seen a declaration 'extern T id1' where T is id2's type just before the #pragma line.  This has the additional benefit that type mismatches between id1 and id2 are caught.
The motivation for this change was to implement the __weak_reference() as
found on many systems.  Usually the alias created by this macro isn't
declared additionally.  Now this macro works:

#define	STR(x)	#x
#define	__weak_reference(sym, alias)	_Pragma(STR(weak alias = sym))

I checked the compiler documentation from HP, Sun and SGI and they appear to
not require a declaration of id1 neither.

Reviewed by:	bp