In the declaration
  class C : <base-classes> {
    ...
  };

access to the base classes must not be done as from within C itself but
as from the enclosing class, function or namespace (11.4#2).  This makes a
difference if a base class B is a private nested class and C is a
friend of B's enclosing class (we allowed access to B although we shouldn't) or
if C's enclosing class is a friend of B's enclosing class (we denied access to
B although we shouldn't).