Fix a delicate issue with friend declarations.  Apparently this was changed rather late in the standardisation process, so the original C++ producer authors probably used the old specification.
If in a declaration

  class A {
      friend class B;
      friend void c();
  };

the class B or the function c were not declared before, their identifiers are
added to the innermost enclosing namespace of class A.  Until here we did
everything right.  This injection caused pain and was very much disliked, so
another rule was added which says that the injected name was not visible
until a further declaration (it is "not found by simple name lookup" in
standardese).

To implement this, I've added a 'invisible' flag to the identifier.  If a type
or a function was first declared in a friend declaration like above, the flag
is set.  On a proper redeclaration it is cleared.  If an identifier is used
with the flag set, an error is emitted.  The flag is also used to exclude
functions from the candidates for overloading.

The invisible flag only needs a single bit.  I had to add a new field because
in 'storage' all bits were already used.  If we need another bit later, we can
turn the invisible field to a second field for various flags.