This is a start on implementing posix3 for #64.
I've been sticking to only those files which haven't changed from posix2 - I think plenty remain, but I'm committing this as-is, so other people can help finish this off, too.

After the rest of these files which simply inherit their posix2 counterparts are done, the next thing to do would probably be enums, which're fairly simple.

Note that I've purposelfully not added this to the Makefile and RUN script yet, as it's incomplete.

svn path=/tendra/trunk/; revision=854