Applied santugin's patch for sid's missing terminal in #68.
I've looked for similar occurances of the same thing elsewhere, and haven't
been able to find any. Do let me know if i've missed them (which is possible,
as the grammar is most confusing!).

It looks like the Makefile for sid dosen't have a rule to regenerate the
parser for sid itself. I realise that having this done automatically and
removing parser.c from the repository would be a catch-22, but having a rule
there to run by hand is handy.

For the moment, though, I'm ignoring that (and I've regenerated parser.c and
junk.h by hand), as there are probably plenty of other similar cases, and I'm
sticking to one thing at a time.

svn path=/tendra/trunk/; revision=859