This is my initial attempt to seperate lexi from TenDRA proper.
The plan is to branch a stable release for backwards compatibility
once the file layout is presentable, and then to clean up features
after that is safely archived.

For the moment, I've simply moved out the relevant parts of the
source as-is; I'm being careful to leave a paper trail which can
be easily followed back to the origional source.

This commit probably breaks building /tendra, but I cannot build
it anyway.

svn path=/lexi/trunk/; revision=1009