Added assertions to check that the formatting strings contain grouping brackets where that is relied upon. These are assertions rather than runtime errors since the formatting strings are hard-coded.
This has to be the most convoluted string-parsing code I've seen; we should be ashamed to host this. Thanks to mauke and maide of ##C for spending the time to  guide me through this.

I've regenerated tdf.c by hand; eventually this should be removed from the repository and generated during each build.

svn path=/tendra/trunk/; revision=1184