Lexi now maintains its own internal fixed-length buffer when reading tokens. This obsoletes the unread_char() function, which is now no longer called by lexi.
This change aims to be backwards-compatible; it should not affect existing programs, though their unread_char() functions may now be redundant and can be removed if not used by their own routines. The examples have been updated accordingly.

The length of this buffer is calculated during the generation of the parser; it need only be long enough to hold the maximum token and mapping lengths. A command line option has been added to disable the generation of assert()s to check this condition, which are enabled by default.

svn path=/lexi/trunk/; revision=1380