This is an initial port of SID's .act handcrafted lexer to a Lexi-generated lexer instead. This has the specific goal of proving zones against a relatively complex case for #167.
Since the goal is so specific, I've diciplined myself in several ways so as not to become distracted:

 * Avoid using Lexi constructs which will be reworked shortly, such as the keyword generation.
 * Refrain from refactoring repeated code if it will be obsoleted by future Lexi features anyway, such as support for identifiers.
 * Work-around non-zone related problems, such as #172.
 * Do not modify the rest of SID's code: this is a strict drop-in replacement (the one exception is adding new error messages).

Hence there are lots of TODOs and XXXs: they are all unrelated to trialing zones, and so will be addressed when this code makes it way over to trunk/ instead.

This is an initial commit, and is simply an "it compiles!" waypoint. The runtime is untested. I am committing now because this is large and complex, and I would prefer bugfix commits have clearly understandable diffs. Do not expect it to run, yet.


svn path=/sid/branches/sid-zonedlexi/; revision=1491