Removed amd64 code.
The code in question was a direct copy of the 80x86 installer and over the
years saw virtually no development towards actually supporting the x86-64
architecture. If someone wants to take this up again in future then they can
always use the latest 80x86 code as a starting point.

Requested by kate and tobiasu.

svn path=/tendra/trunk/; revision=1626