This is a start on moving the filetype-related tables of data into one central place. These were mostly in the form of macros defined in various headers which were expected to be kept in sync. This first pass on refactoring sees these collated into one array, and a vestigial API drawn from previous operations.
To follow are the actions associated with the filetypes, which are currently nests of special-case actions scattered elsewhere. The requirements of these actions should form the majority of the API for `table.h`, allowing its internals to be kept private.

Where appropiate, a start has been made on moving enumerated types to use `enum` rather than unrestricted integers.

svn path=/trunk/; revision=2079