Assigned `sid_current_entry` when creating actions. This provides the action name when giving action-related error messages further down the parse, in `<x-action>`. Regenerated accordingly.
This was the cause for a few relatively obscure segfaults, which boiled down to `&entry->key` to find the key name for the current entry (which is used to give the name of the action to which the error pertains, which would have incorrectly stated the previous entry encountered). Since it was possible to reach this before any other entries had been formed, that took the address relative to a `NULL` pointer.

svn path=/trunk/; revision=2142