A start on simplifying the document processing. This changeset brings together the Docbook sanitization from Doctools along with TenDRA's customisations, plus CSS structured for the forthcoming Trac reorganisation for the new website.
Doctools will be removed as a dependency once the existing documents are converted to be build using a central makefile facility for documentation.

The net result is that everything ought to be in one place, have just one way to do things, and to be structured as simply as it can get.

svn path=/trunk/; revision=2275