Added support for NetBSD 5/80x86 for the following APIs: ANSI C89, ISO C90, POSIX2, POSIX2, XPG3, XPG4, and miscellany. Not supported are SVID3 and UNIX95, which NetBSD appears to not attempt to provide.
This is a rework of machines/netbsd/80x86/{include,startup} from scratch, in an effort to lay a good template for future implementations to follow. In particular, I have taken care to keep modifications for hacked includes to contain their workarounds with per-version guards, using the version macro (_NETBSD5) provided by makedefs.

svn path=/trunk/; revision=2286