Split off `<tendra.obj.mk>` from library and program building. This commit sees the suffix rules for `.o` files removed.
As a side-effect of moving to `<tendra.obj.mk>`, `.PATH` searching in the producers is also removed. This folds in all sources into a flat space; for the moment, these are explicitly specified. However, now use of suffix rules (and therefore `.PATH`) is gone, I have grouped these per directory, with the intention of replacing these with partial linking per directory. This is another of those tedious commits which sets the stage for future refactoring.

svn path=/trunk/; revision=2300