Refactoring. This builds on the extensible printf facility provided in [2195] to implement application-specific formatting specifiers catering for the error message reworking outlines in the [wiki:2008/07/19/16.50 recent post on overengineering].
This sees the `.e` files go (and thus the `errors/` directory removed), as each error message is translated to use [source:/trunk/tendra/src/shared/error/error.h shared/error] instead.

As a side-effect, Perl is no longer a dependancy for building TLD now that the `.e` files are gone.

svn path=/trunk/; revision=2327