Moved out startup files (modes, portability tables and such, not the API building startup files, which unfortunately have the same name).
I was a little unsure as to whether these belong with the producers, or with tcc. Certianly they configure the producers, but then so do all the options tcc passes. I've opted to put them with tcc for two reasons: firstly to keep all tcc's configuration stuff in one place (alongside the environments), and secondly because if these were placed with the producers, you'd still need tcc to find them (unless you went exploring the filesystem by hand). In other words, they're here because this is where they're used.

svn path=/trunk/; revision=2375