Remove the `makedefs` script, and its generated `Makefile.defs` file. Instead, the variables this produced are calculated on-the-fly by the various assignments in `<tendra.makedefs.mk>`.
The idea here is that these variables are only assigned if required, and that the majority of projects do not use them (as they typically do not require knowledge of any particular machine). In this way we attempt to avoid inconveniencing the user with running a `makedefs` script per-project.

Since no `Makefile.defs` is generates, this also keeps the source tree read only.

svn path=/trunk/; revision=2384