Move out the C++ LPI token implementations and the C++ (minimal) standard library to the producer project.
I am unsure of the exact dependency interactions here, and I am not certian that this is where this code will finally live. I am also undecided as to whether these are specific to the DRA producers, or applicable to all C++ LPI interfaces in general. In any case, these are currently disabled from the build (which means compiled C++ will be missing tokens, of course).

My current goal is to get top-level bootstrap working, and then to re-integrate these once the rest of [source:/trunk/tendra@2417 /tendra] is gone.

svn path=/trunk/; revision=2422