Specify the default environment by name, rather than by absolute path. The environment path will be searched, and is now pre-populated by what used to the the absolute path to the installed default environment.
Making use of the path-searching mechanism here permits calling an installed `tcc(1)` with `$TCCENV` set to search not-yet-installed environments. That feature is required since environments now live elsewhere than the tcc code, along with the API compilation, but tcc is depended upon for that compilation.

My apologies for calling tcc's option-parser directly here - there doesn't seem to be a user-facing way to specify environment paths, and I didn't want to add one. The more typical interface would be to set `$TCCENV`, but I found that to be overkill for this case.

svn path=/trunk/; revision=2427