Quite a few things happening at once, here. The crux of this commit is that the trans(1) manpages are now converted to Docbook.
These are all very similar, differing only in a few details. So, I have consolidated them into one file, an XSLT stylesheet, which may be configured to render out whichever is appropiate. It can also render out a consolidated page, giving usage for all translators.

The infrastructure for dealing with multiple documents in the Makefiles isn't the focus of this commit; if we do gain support for building multiple files simultaneously, these can be generated on-the-fly at that point. For the moment, I am simply committing the generated source.

svn path=/trunk/; revision=2440