Moved out `libtdf` into its own project, now known as `libtsl`. This is the last of the code from `trunk/tendra`, which can now go.
I've made `libtsl` a standalone project (rather than including it with the installers) for two reasons: firstly, build convenience (as it depends on complete end-to-end C compilation), and secondly because it is not specific to the DRA installers.

This commit just serves to get things in the right place; the tcc environments and suchlike need updating accordingly. That'll come along with the bootstrap makefile, which is what's driving this project-by-project systematic sorting-out of things.

svn path=/trunk/; revision=2469