Some normalisation for makefile variables, now I'm done moving things around; this marks the start of clearing up the post-restructuring aftermath.
The various `*_DIR` variables have been replaced with `PREFIX_*` instead, and used more consistently. Since the layout under `$PREFIX` is now per-project, the `machines` directory is gone, and so this commit removes variables associated with that.

Hopefully this should be a bit simpler for package maintainers to configure, by overriding whatever they wish.

svn path=/trunk/; revision=2493