Regeneration for Blueprint css.
This is to be used by both the standalone documents (installed under `$PREFIX_HTMLDOC`) and by the website's pages, for consistency between both. Hence it cannot be generated on-the-fly during the website build, because that is not present at the time documents are rendered for standalone use.

I didn't want to introduce ruby as a dependency for packages, so I'm committing this by hand, as generated source.

This is a start on centralising a few remaining bits and pieces which need bringing together between the website and the standalone documentation.

svn path=/trunk/; revision=2551