Typo. Hunted down by Tom Beebe and Robert Smith.
This was part of the old perl-generated error reporting from libexds, replaced for TLD in [2327]. Unfortunately the replacement mistakenly marked this particular error at the wrong severity.

Before that refactoring, TLD relied on the default minimum severity level to elide warnings, including this one. This warning is present for typical use (to complain about tokens which are not present, because they are provided by libc or equivalents). So, it is not something we'd want to bother the user with, during a typical run.

For now, I'm simply changing the severity to `ERROR_WARNING`. This will be rather verbose for the user, but lets the build continue. The verbosity is essentially unrelated and can be addressed in a later commit.

svn path=/trunk/; revision=2554