Import of the "Third Eye" symbol table definition. This is the origin to the similar headers found in many projects (notably gdb and gas).
mipstrans expected this to be present under `/usr/include` on a user's machine. However, my current aim is to have the installers compilable on any architecture, regardless of the code they generate (i.e. to build them as cross compilers). Hence it is pertinent to remove that dependency.

The copies present in the gdb/gas source (as far as I gather) have been reworked to depend on GNU libbfd. I'd rather TenDRA didn't depend on that, so I'm using this older version, which I believe is closer to what was used by the origional developers.

I believe this file is unmodified as of when it was published for general use, as evidenced by it being copied into many opensource projects.
See [http://datahedron.com/mips.html] for some history.

svn path=/trunk/; revision=2584