Ported to OpenBSD/sparc.
This is a combination of the OpenBSD/80x86 environments, startup files and hacked includes, and the Solaris/sparc TDF tokens and installer.

My intent here is primarily to target platforms for which support can be pieced together from existing code, so that I can expose potential problems in the porting process, and also to help understand what can and cannot be centralised when refactoring the layout of `/osdep`.

The one issue I found here is that the installer generates `fmovd` instructions. These are not supported by SPARC v8, but do exist in v9. In Solaris, `fmovd` is trapped by the kernel and emulated at runtime; I don't know if OpenBSD does the same.

What I'd really like is to add a `-K` flag to specify which set of instructions to produce, and have this emulation done explicitly in the generated code. Meanwhile, I've just set the environment to pass `-Av8plus` to GNU as(1), which presumes this instruction will be trapped.

Tested on OpenBSD 5.0 on a SPARCStation 5.

svn path=/trunk/; revision=2595