Restructuring for `shared/src`. This presents headers visible as: {{{ #!c #include <shared/check.h> #include <shared/error.h> }}}
The various shared facilities are now linked as one archive, because (almost) all projects use all of them anyway. That won't significantly change for future additions, as more of the codebase gets centralised here.

I'm viewing this change as representative of something more fundamental; I'm thinking I can keep the code layout for shared things (in their native location) in whatever manner best suits there - and that the individual uses of this per-project need not neccessarily have the same layout. So I'm using symlinks to cherry-pick things, to present those in a directory structure which suits each project.

svn path=/trunk/; revision=2621