Bugfix for dereferencing freed tuples.
There are two sets of tuples (parameters and results) and two main paths through the code here.

When redefining entries, these tuples are destroyed and their local counterparts used instead. However, the checks introduced in [2137] reference the (now destroyed) tuples; this commit restructures this to use their local counterparts instead.

I hope this is correct, because it's a little inconvenient to test. The only existing grammar I've found which triggers this path (for results, not params) is the dra-producer's C++ grammar.

svn path=/trunk/; revision=2626