Remove argument ordering and the `-no_shuffle` flag. Those were introduced by Amos in [363].
As far as I understand, the motivation behind ordering arguments was to permit the (newly-added) `-y` options to be given after `-Y` environments, and to still have effect on the loaded environment. That mechanism is now provided by precidence within the hash of environment variables instead, and I see no other reason to keep ordered nodes.

svn path=/trunk/; revision=2648