Moved out the non OS-dependent tcc environments under `tcc/env`. These are the bare minimum for running tcc itself for a few situations (notably with `-Ft` and `-Fj`) without involing the LPI stages for producers and beyond.
There's a bit of a hack here for the Solaris `X*` environments; Solaris provides `/lib/values-Xa.o` and friends which correspond to some of the environments for tcc. Rather than view the `X*` environments as non-portable just because of this exceptional case, I'm having the osdep environments overwrite their `tcc/env` equivalents.

This work is done to allow `${TCC_IN_SITU}` to be removed from `osdep/` - however, running tcc "in situ" will be required at a later point for building the DRA producers C++ LPI, so I am not removing `tendra.situ.mk` just yet.

svn path=/trunk/; revision=2651