Various miscellaneous fixes for compiler strictness. These roughly fall into the following categories:  * Removing unused variables and functions. Some are conditional by preprocessor macros;  * Missing headers;  * Implicit types;  * Braces for nested conditional blocks;  * Rephrasing for unreachable code;  * Missing comment terminators. These are a side-effect of the copyright reworking.
In particular, warnings about possible use of uninitialised values are ignored, for two reasons; firstly (the installers in particular) make use of passing structs by value, which often have partially-populated fields, read conditonally. We ought to deal with these in a less brutal way than simply initialising them unneccessarily. Secondly, static analysis is often unable to cope with conditional code paths; in these situations, the code ought to be reworked for simplicity. So, I have left those untouched, and they will still warn, if a compiler checks for that.

Regenerated.

svn path=/trunk/; revision=2704