Removed compilation scheme diagrams.
These seemed excessive here; two are duplicates taken verbatim from the TDF and Portability paper (papers/porting) and one appears to be a faithfully recreated almost identical copy, save for the addition of an assembly stage, which I am adding to the diagrams in that paper.

Instead, I've worded the document to urge the reader to consult that paper. This sees the "TDF Compilation Strategy" chapter go entirely.

svn path=/trunk/; revision=2706