Oops... I missed a lot of things I'd marked `__WRONG` earlier, when providing support for GLIBC 2.14.
In this commit I've re-enabled those (and made a reasonable attempt to provide hacked includes, which I'll commit seperately).
More importantly for this commit, I've made an attempt to clean up the startup headers somewhat; in particular the collection of `#pragma` workarounds in `ansi.h` now live in the hacked includes, as per each specific use.

I'm trying to keep the startup headers to only define the feature-selection macros specified by each standard. In some cases this actually disables support for various things within that API. But I think this is preferable than enabling non-standard things here; that is better done in the hacked includes instead. So keep it clean in startup land, please.

svn path=/trunk/; revision=2725