Centralise the feature-selection macros used to configure the system headers per API, during TenDRA's API building phase.
These are not OS-specific, and so they are moved out of the osdep/ code, in order to reduce repetition. They now live with tspec, alongside their respective APIs.

I am not confident that all of these are correct, but I've done my best for the moment. Each feature/$api.h header here should define the feature-selection macros required for that API. These are picked up by the system headers, and make visible relevant interfaces.

During API building, these are included before the OS-specific hacks, so that those hacks may `#undef` or override things if neccessary. So feature/ represents an ideal world, of what the standards expect.

svn path=/trunk/; revision=2732