A (long overdue) rewrite of the tcc environment file parser. This is implemented as a lexer (written using Lexi), and a parser (which is just a simple loop, as the grammar is trivial).
The main purpose of this is to split off the environment parsing from the options parser, which facilitates various work on the parsing. Most of the work here is done during lexing - for example, that's where the `<ABC>` substitutions take place.

I've kept compatibility with the existing files except for a couple of minor points; namely that `+XYZ` used to append in some cases, depending on the variable name. Now the +/</> operators are consistent no matter what the variable.

This breaks the build due to the dual location of information in tcc, which is due to be removed in the next commit.

svn path=/trunk/; revision=2734