Removed `GCC_STRUCTS`, and moved `remove_struct_ref` out of config.h.
`gcc_compatible` is now a command line option, -t. This used to be -G, but that was incorrect; -G is to be used for the assembler dialect, and `gcc_compatible` represents ABI control.
 
I am unsure about some of this, but I'll work through it after the dust settles and the OS-specific directories are gone. For the moment I'm just moving things out of the way. The -s and -t flags will eventually be removed.

svn path=/trunk/; revision=2801