Presented the svsV_assembler flag as `-b` through the CLI. This removes its dependence on the `SYSV_AS` macro, which is now removed.
This appears to be a combination of assembler dialect and ABI for the sake of producing a.out vs. ELF assembly. I'm keeping it around for the moment until I can test exactly what needs to happen for a.out systems.

Forthcoming work will split this into ABI and as(1) dialect things, or have a seperate flag for executable format.


svn path=/trunk/; revision=2809