Executable format is now a command line option, -F, provided for all architectures. This removes the -k option for 80x86.
For most architectures we currently only provide one format, but more will come (notablly ELF) with support for other OSes on the existing architectures.

This is the groundwork for moving out each architecture's "machine.c" and friends, which is responsible for generating format-specific things such as switching assembler sections. My intention is to share this format-specific code between not only all OSes of one architecture, but between all architectures.

svn path=/trunk/; revision=2820